20221219-01(Buff全)

+++ Assets/Script/SO_Player.cs
@@ -11,7 +11,7 @@ public class SO_Player : ScriptableObject
     public float Hp;
     public float Mp;
     public float MpUp;
-    public int Atk;
+    public float Atk;
     public int Def;
     public int Rec;
